MetLife/requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
colorlog==4.4.0
configparser==5.0.0
contextvars==2.4
crayons==0.4.0
cryptography==3.1.1
et-xmlfile==1.0.1
fabric==2.5.0
fake==0.8
Faker==4.9.0
filetype==1.0.7
har2case==0.3.1
httprunner==2.5.7
id-validator==1.0.12
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
invoke==1.4.1
jdcal==1.4.1
Jinja2==2.11.2
jsonpath==0.82
jsonschema==3.2.0
loguru==0.5.1
MarkupSafe==1.1.1
more-itertools==8.4.0
MouseInfo==0.1.3
numpy==1.19.2
openpyxl==3.0.4
packaging==20.4
paramiko==2.7.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
PyAutoGUI==0.9.50
pycparser==2.20
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pymssql==2.1.4
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pytest==5.4.3
python-dateutil==2.8.1
PyTweening==1.0.3
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
sentry-sdk==0.13.5
six==1.15.0
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.2.5
webdriver-manager==3.2.2
win32-setctime==1.0.2
xeger==0.3.5
zipp==3.1.0
